Separate out cocos specific manual type conversions.
