+++ scripting/javascript/bindings/cocosjs_manual_conversions.h
@@ -1,0 +1,19 @@
+#include "chipmunk.h"
+#include "cocos2d.h"
+#include "js_manual_conversions.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+  
+JSBool jsval_to_CGPoint( JSContext *cx, jsval vp, cpVect *out );
+jsval CGPoint_to_jsval( JSContext *cx, cpVect p );
+JSBool jsval_to_cpBB( JSContext *cx, jsval vp, cpBB *ret );
+jsval cpBB_to_jsval(JSContext *cx, cpBB bb );
+    
+#ifdef __cplusplus
+}
+#endif
+
+#define cpVect_to_jsval CGPoint_to_jsval
+#define jsval_to_cpVect jsval_to_CGPoint
