+++ scripting/javascript/bindings/js_bindings_chipmunk_functions.cpp
@@ -5,7 +5,7 @@
 
 #include "jsfriendapi.h"
 #include "ScriptingCore.h"
-#include "js_manual_conversions.h"
+#include "cocosjs_manual_conversions.h"
 #include "js_bindings_config.h"
 #include "js_bindings_chipmunk_functions.hpp"
